Change X, Check fill color
